working without optional Classes as inputs
