+++ base/Cargo.toml
@@ -7,15 +7,3 @@ edition = "2021"
 
 [dependencies]
 thiserror = "*"
-uniffi = { version = "*", features = ["cli"] }
-
-[build-dependencies]
-uniffi = { version = "*", features = ["build"] }
-
-[lib]
-name = "base"
-crate-type = ["lib", "cdylib", "staticlib"]
-
-[[bin]]
-name = "uniffi-bindgen"
-path = "uniffi-bindgen.rs"
