+++ base/src/lib.uniffi.rs
@@ -1,0 +1,8 @@
+mod error;
+mod module;
+
+pub use error::Error;
+use module::{Class, Uniffi, WrappingClass};
+
+uniffi::include_scaffolding!("base");
+
