+++ base/uniffi/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "base_uniffi"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+thiserror = "*"
+uniffi = { version = "*", features = ["cli"] }
+
+[build-dependencies]
+uniffi = { version = "*", features = ["build"] }
+
+[lib]
+name = "base_uniffi"
+path = "../src/lib.uniffi.rs"
+crate-type = ["lib", "cdylib", "staticlib"]
+
+[[bin]]
+name = "uniffi-bindgen"
+path = "uniffi-bindgen.rs"
